Contenidos/05_ia_ds_modulo_5/01_ejercicios/03_semana_3/app_con_ia/app/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
httpx
markovify
openai
python-dotenv
google-generativeai